Set io.netty.native.workdir on application master (#8)
* Fix for non-executable tmp directory

This changes the docker setup to set `noexec` for `/tmp`, which is true
for some systems. The `ApplicationMaster` is fixed to work on such
systems. For some reason the system property wasn't being picked up via
the command line `-Dio.netty.native.workdir` flag, so we specify it in
the application master code. This is fine, and shouldn't be a problem.