Support robust object cleanup
Add `with_finalizers` decorator. This provides support for robust object
finalization in both python 2 (complicated) and python 3 (much easier).
The guarantees are:

- Cleanup functions are always called at *some* point on non-fatal
  shutdown of python interpreter.

- In Python 3, cleanup functions are always called when the object loses
  its last reference. In the case of direct reference cycles, this
  happens after a full gc run.

- In Python 2, cleanup functions are run when the object loses its last
  reference, unless its part of a direct reference cycle. In that case
  cleanup doesn't happen until interpreter shutdown.

In both cases, "direct" means the object itself has a reference that
eventually leads back to itself. This is opposed to reference cycles
that happen to also contain a (non-cyclic) reference to the object.

```python
a.b = b
b.a = a

b.c = c
c.b = b
b.a = a  # referred to by cycle, but not part of cycle
```