Test user permissions in containers (#22)
* Test user permissions in containers

Previously run containers with *simple* authentication wouldn't be able
to access hdfs as the user. This is now fixed by setting the
`HADOOP_USER_NAME` environment variable in each container. Note that
`whoami` will still report as `yarn` (or whatever the yarn nodemanager
was started as) under *simple* authentication. As far as I can tell,
the only way to get around this would be to change the login at a
service level, I don't think there's a way to get yarn to start a
container as a user when using *simple* authentication.

Of course all of this already worked fine for kerberos, showing that
even when something is called "simple" with hadoop security, it's not
really "simple".

---

Also, use the `finalize` implementation from `backports.weakref` on
python 2 instead of rolling our own.

* Update travis setup

* Make tests more robust