Ensure consistent ownership state
Previously the following steps would result in inconsistent ownership
state in the key-value store:

- Create a key
- Assign an owner to that key
- Delete the key
- Create the key again, without an owner

At this point the key-value store would say the key was unowned, but the
``ServiceTracker`` would still have the key in its set of owned keys.

We now test for this case, and owned keys from their service trackers
upon deletion.