Added support for ViewFs
Prior to this commit the Daemon only accquired a delegation token for
defaultFs which did not work well in the presence of federation. Now the
application spec contains a list the namenodes to acquire the delegation
token for in addition to defaultFs. In theory, all tokens should be
automatically renewed by the RM. However, I did not have the occasion
to test this.