Add a Web UI (#68)
Adds a web-based UI for monitoring the application.
The states of all services and the key-value store are exposed.
As per the YARN security documentation, the webui is secured
using the AmIpFilter, and has been tested locally. Smoketests
have been added for the UI, but most testing has to be manual
for now.