Error nicely if not logged in
Previously the daemon server would hang if the user hadn't kinit'd
before running an operation. This is due to `UserGroupInformation` not
caring if no credentials were found on init and only checking when an
operation is attempted, coupled with the errors being logged only (no
way to catch failed login at call time, the yarnclient just hangs).

We now manually check if credentials are missing before doing any
operation that might hang, and error accordingly. Once
`UserGroupInformation` has been init'd, there's no way to reset (it will
never look for a kerberos ticket again), the daemon has to be restarted.

This is annoying.

To make sure the user clients always get a nice error message (instead
of a failed connection) we hang around and respond only with errors
saying to "kinit and restart the daemon" to all requests until shutdown.