Use NM_HOST to determine hostname
Some clusters have infiniband enabled, which may get picked up by
`InetAddress` when determining the application-master hostname/IP.
Infiniband is not currently supported by gRPC, so this would result in
failures. Instead of trying to more robustly determine the IP for a
supported network interface, we rely on the `NM_HOST` environment
variable, which is always set and must resolve correctly on all nodes
for hadoop to properly work.