Make security configurable per application
Allows specifying security configuration as part of the application
spec. TLS credentials can be specified either as file locations or as
raw bytestrings. This allows for launching applications with differing
TLS credentials from the same originating daemon.

Also quiets a noisy extraneous log message from grpc >= 1.15 when using
fork + exec. This noisy extraneous message should be fixed in a future
release of grpc (1.17?), but for now it's better to just quiet it
ourselves.