Rename "Daemon" to "Driver"
The "Daemon" process name has been confusing for some users, here we
rename it to "Driver" to better reflect its purpose and usage.

Note that this is (currently) a backwards incompatible change for expert
users that may have been touching lower-level methods/classes. This will
be fixed in a subsequent commit deprecating those features.