Rename daemon to driver (#116)
* Rename "Daemon" to "Driver"

The "Daemon" process name has been confusing for some users, here we
rename it to "Driver" to better reflect its purpose and usage.

Note that this is (currently) a backwards incompatible change for expert
users that may have been touching lower-level methods/classes. This will
be fixed in a subsequent commit deprecating those features.

* Deprecate daemon smoothly for driver

Raise nice warnings when old `Daemon` commands/classes are used. This
includes:

- `DaemonError`
- `DaemonNotRunningError`
- `skein daemon` cli command
- `Client.start_global_daemon`
- `Client.stop_global_daemon`

* Update changelog [skip-tests]