Add ability to run user code on the application master
Allows running a single user process on the application master. This
allows singleton services to start up *much* faster.

The master process (currently called a "driver", but that's a bit too
overloaded for my liking) also manages the lifetime of the application -
when the process exits the application shuts down automatically,
regardless of the state of other services. The application final status
is determined by the exit code of the process (0 for succeeded, anything
else for failed).

This can be useful for running things like the dask client for
dask-submit (more robustly ensure the cluster shutsdown after
submission) or starting single processes like a jupyter notebook server
(used in jcrist/yarnspawner to launch jupyterhub on YARN).