Add ``skein driver stop --force``
Previously the global driver would only be stopped if skein could verify
that the corresponding PID was actually a skein driver. In the case of
credential mismatch this could result in an inability to stop a running
driver using the skein cli. This adds a `--force` option to always kill
the process, even if we're not sure if it's a driver.