Allow creating Client in kerberized container
Previous work was done to enable submitting applications from inside a
YARN cluster container, but this failed for kerberized clusters. We now
grab a delegation token for the resource manager to add to the set of
credentials distributed throughout the application. This allows a
YarnClient to be created and used in all containers managed by Skein.