Don't match on vcores for allocated containers (#131)
Some YARN configurations return ``Resource`` instances on allocation
that have vcores always set to 1 (even though the allocated container
has access to the requested number of vcores). This led to skein not
matching the allocated resource to any containers when the requested
vcores > 1. Now we ignore vcores on matching, and use only the requested
vcores throughout.