Refactor ApplicationMaster shutdown handling (#134)
This PR refactors the shutdown handling code for the application master,
simplifying it and organizing it in a way that reduces the number of
potential code paths. This fixes a periodic test failure that turned up
some potential issues when the application master is killed due to a
memory error (not correctness issues, but the logs weren't always useful
for debugging).

While refactoring this I also came across a very rare race condition in
the allocator loop (a container could be requested, canceled, and then
still received if one code path was taken). This has also been fixed.