Add `allow_failures` field (#145)
* Add `allow_failures` field

Adds support for allowing failures in a service. If set to True, the
application will not terminate if the number of failures for a service
exceeds `max_restarts` for that service. Default is False.

This is useful for things like `dask` workers, where we may want to
handle restart behavior manually instead of relying on skein to do it
for us.