Fix skein.properties.container_dir
Fixes a bug (introduced in 0.6.0) in detecting the absolute path to the
container directory.  YARN (at least some versions) will create all
possible container directories (as listed in
`yarn.nodemanager.local-dirs`, but only populate one of them. Previously
we were just looking for a directory to exist when searching for the
container directory, now we look for the proper files as well.