Fix shutdown condition (#173)
* Fix shutdown condition

In the absence of restarts, a failing container should cause the whole
application to fail. However, the code implementing this was broken.
This fixes that and adds tests to enforce the specified behavior.

* xfail flaky memory_limit test

The shutdown order of an application master killed with a memory error
isn't deterministic - this can lead to the test periodically failing.
Xfail it for now.