Add `ApplicationClient.add_container` (#174)
* Add `ApplicationClient.add_container`

Adds a method to add a new container to a service, with optional
service configuration overrides. So far only environment variables can
be overridden, additional fields could be added as needed.

    # Add a new container for service `test`, with additional
    # environment variable `FOO` set to `BAR` in the container
    app.add_container('test', env={'FOO': 'BAR'})

This is useful for services where every container should be *almost*
identical, with small changes made to update configuration parameters as
needed. I needed this for ``dask-gateway``, where each worker needs a
unique name provided to it before startup.

In the presence of container restarts, the container is restarted with
the same overridden configuration. A test is added to check this
behavior.

* Flaky get_applications test

Only compare application names, the resource manager doesn't update all
fields atomically leading to potential differences in state between
requests.